Add Genre and GenresCategories entities and configurations
Implemented new entities and configurations for `Genre` and `GenresCategories` in the `CodeflixCatalogDbContext`. Created `GenreConfiguration` and `GenresCategoriesConfiguration` classes to define their entity configurations. Introduced a `GenresCategories` class for the many-to-many relationship between `Genre` and `Category`. Updated the `GenreRepository` to support the insertion of `Genre` entities with their associated categories. Added tests in `GenreRepositoryTest` to verify the correct establishment of relationships and enhanced the `GenreRepositoryTestFixture` for generating example entities.